Fix issue with FileTransferClient::MAX_FileChunk. This is not java, meaning that you cannot declare and define a const variable in the class definition, which means you have to define storage for it in the cpp file. Unfortunately, MAX_FileChunk is used for array storage in the class definition.
Use a #define for now, and cast it to unsigned short where neccesary.
Observe the compiler warning (not introduced with this commit) though:

CVSnsferClient.cpp:214: warning: comparison is always false due to limited
   range of data type

Which of course is true. I don't know the protocol specifications on the
top of my head, but you might be able to change the offending
statements to use int instead.

At least it's usable now.